Feature: Exclude columns
Add command line option 'exclude', which allows the user to specify a comma
separated list of column values to exclude from writing back to file.
For example if there are correct values in the google sheet and
incorrect values in the source flatfile format.

Additionally adjust the style of the source code files to yapf
standards.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>